requirements.txt
boto3==1.21.29
botocore==1.24.29
crispy-bootstrap5==0.6
Django==4.0.4
django-ckeditor==6.2.0
django-countries==7.3.2
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-storages==1.12.3
gunicorn==20.1.0
Pillow==9.0.1
psycopg2==2.9.3
whitenoise==6.0.0